migrate:undo-integer-to-float

+++ src/app/models/Plan.js
@@ -6,7 +6,7 @@ class Plan extends Model {
       {
         title: Sequelize.STRING,
         duration: Sequelize.INTEGER,
-        price: Sequelize.INTEGER,
+        price: Sequelize.FLOAT,
       },
       {
         sequelize,
